requirements.txt
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.4.22
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
python-dotenv==1.2.2
python-telegram-bot==22.7
tzdata==2026.2
tzlocal==5.3.1
